% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sex.R
\name{get_sex}
\alias{get_sex}
\title{Determine le sexe d'un individu}
\usage{
get_sex(name, birth_date = NULL, birth_year = NULL, age = NULL)
}
\arguments{
\item{name}{Un prenom. Non sensible a la casse sur les majuscules.}

\item{birth_date}{Date de naissance, au format Date.}

\item{age}{Age en années.}

\item{birth_date}{Annee de naissance.}
}
\value{
Le sexe estime de l'individu, 'F' ou 'M'.
}
\description{
Determine le sexe d'un individu
}
\details{
Renseigner aucun ou un seul des trois arguments relatifs a la date de naissance.
Si aucun n'est renseigne, le sexe est determine sur la periode 1900-2017 entiere.
}
\examples{
get_sex('Robin', birth_year = 1995)
get_sex('JEAN-JACQUES', age = 59)

}
