% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prenoms_nat.R
\docType{data}
\name{prenoms_nat}
\alias{prenoms_nat}
\title{Prenoms des Bebes Francais 1900-2018 National}
\format{
Un data frame avec 601 223 lignes et 5 colonnes :
\itemize{
  \item{\code{year} : année (1900-2018)}
  \item{\code{name} : prénom}
  \item{\code{sex} : sexe (M/F)}
  \item{\code{n} : nombre de nouveaux nés}
  \item{\code{prop} : proportion de nouveaux nés de ce sexe portants ce prénom cette année}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/2540004}
}
\usage{
prenoms_nat
}
\description{
Prénoms des nouveaux nés Français entre 1900 et 2018, au niveau national.
}
\keyword{datasets}
