% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_sex.R
\name{estim_sex}
\alias{estim_sex}
\title{Determine le sexe d'un individu}
\usage{
estim_sex(name, birth_year = NULL, age = NULL, birth_date = NULL)
}
\arguments{
\item{name}{Un prenom. Non sensible a la casse.}

\item{birth_year}{Annee de naissance.}

\item{age}{Age, en annees.}

\item{birth_date}{Date de naissance, au format Date.}
}
\value{
Le sexe estime de l'individu, 'F' ou 'M'.
}
\description{
Determine le sexe d'un individu
}
\details{
Renseigner aucun ou un seul des trois arguments relatifs a la date de naissance.
Si aucun n'est renseigne, le sexe est determine sur la periode 1900-2018.
}
\examples{
estim_sex('Robin', birth_year = 1995)
estim_sex('JEAN-JACQUES', age = 59)

}
