% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prenoms_dpt.R
\docType{data}
\name{prenoms_dpt}
\alias{prenoms_dpt}
\title{Prenoms des Bebes Francais 1900-2018 Departemental}
\format{
Un data frame avec 3 567 753 lignes et 6 colonnes :
\itemize{
  \item{\code{year} : année (1900-2018)}
  \item{\code{name} : prénom}
  \item{\code{sex} : sexe (M/F)}
  \item{\code{dpt_num} : numéro du département}
  \item{\code{dpt_name} : nom du département}
  \item{\code{n} : nombre de nouveaux nés}

}
}
\source{
\url{https://www.insee.fr/fr/statistiques/2540004}
}
\usage{
prenoms_dpt
}
\description{
Prénoms des nouveaux nés Français entre 1900 et 2018, au niveau départemental.
}
\keyword{datasets}
