% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prenoms_dpt.R
\docType{data}
\name{prenoms_dpt}
\alias{prenoms_dpt}
\title{Prenoms des Bebes Francais 1900-2018 Departemental}
\format{Un data frame avec 3 567 753 lignes et 5 colonnes :
\itemize{
  \item{\code{year} : annee (1900-2018)}
  \item{\code{sex} : sexe, M ou F}
  \item{\code{name} : prenom}
  \item{\code{n} : nombre de nouveaux nes}
  \item{\code{prop} : proportion de nouveaux nes portant ce prenom cette annnee la et dans ce departement}
}}
\source{
\url{https://www.insee.fr/fr/statistiques/2540004}
}
\usage{
prenoms_dpt
}
\description{
Prenoms des nouveaux nes Francais entre 1900 et 2018, au niveau departemental.
}
\keyword{datasets}
