% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prenoms_nat.R
\docType{data}
\name{prenoms_nat}
\alias{prenoms_nat}
\title{Prenoms des Bebes Français 1900-2018}
\format{Un data frame avec 601 223 lignes et 5 colonnes :
\itemize{
  \item{\code{year} : annee (1900-2018)}
  \item{\code{sex} : sexe, M ou F}
  \item{\code{name} : prenom}
  \item{\code{n} : nombre de nouveaux nes}
  \item{\code{prop} : proportion de nouveaux nes portant ce prenom l'annee en question}
}}
\source{
\url{https://www.insee.fr/fr/statistiques/2540004}
}
\usage{
prenoms_nat
}
\description{
Prenoms des nouveaux nes Français entre 1900 et 2018, au niveau national.
}
\keyword{datasets}
